feat(react-native): lobby customization
