+++ packages/react-native-sdk/src/components/Call/CallControls/LobbyControls.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import { StyleSheet, View } from 'react-native';
+import { ToggleAudioPreviewButton } from './ToggleAudioPreviewButton';
+import { ToggleVideoPreviewButton } from './ToggleVideoPreviewButton';
+import { theme } from '../../../theme';
+
+/**
+ * Controls for the Lobby Component
+ */
+export const LobbyControls = () => {
+  return (
+    <View style={styles.container}>
+      <ToggleAudioPreviewButton />
+      <ToggleVideoPreviewButton />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    paddingVertical: theme.padding.md,
+    flexDirection: 'row',
+    justifyContent: 'space-evenly',
+  },
+});
