+++ packages/react-native-sdk/src/components/Call/CallControls/index.tsx
@@ -9,3 +9,4 @@ export * from './ToggleCameraFaceButton';
 export * from './ChatButton';
 export * from './ReactionButton';
 export * from './CallControls';
+export * from './LobbyControls';
