+++ packages/react-native-sdk/src/components/Call/Lobby/index.ts
@@ -1,0 +1,2 @@
+export * from './Lobby';
+export * from './JoinCallButton';
